+++ govtool/frontend/src/components/atoms/VotingPowerChips.test.tsx
@@ -1,28 +1,15 @@
 import { render, screen, fireEvent } from "@testing-library/react";
 import { describe, it, expect, vi } from "vitest";
 import * as Hooks from "@hooks";
-import * as Context from "@context";
 import * as Utils from "@utils";
 import { VotingPowerChips } from "@atoms";
 
 describe("VotingPowerChips", () => {
-  const mockUseCardano = vi.spyOn(Context, "useCardano");
-  const mockUseGetDRepVotingPowerQuery = vi.spyOn(
-    Hooks,
-    "useGetDRepVotingPowerQuery",
-  );
   const mockUseScreenDimension = vi.spyOn(Hooks, "useScreenDimension");
   const mockCorrectAdaFormat = vi.spyOn(Utils, "correctAdaFormat");
   const mockUseTranslation = vi.spyOn(Hooks, "useTranslation");
-  const mockUseGetVoterInfo = vi.spyOn(Hooks, "useGetVoterInfo");
 
   it("renders loading spinner when data is loading", () => {
-    mockUseCardano.mockReturnValue({
-      isEnableLoading: 'demos',
-    } as ReturnType<typeof Context.useCardano>);
-    mockUseGetDRepVotingPowerQuery.mockReturnValue(
-      {} as ReturnType<typeof Hooks.useGetDRepVotingPowerQuery>,
-    );
     mockUseScreenDimension.mockReturnValue({
       isMobile: false,
       screenWidth: 1024,
@@ -30,21 +17,12 @@ describe("VotingPowerChips", () => {
     mockUseTranslation.mockReturnValue({
       t: (key: string) => key,
     } as ReturnType<typeof Hooks.useTranslation>);
-    mockUseGetVoterInfo.mockReturnValue(
-      { voter: { isRegisteredAsDRep: true } } as ReturnType<typeof Hooks.useGetVoterInfo>,
-    );
 
-    render(<VotingPowerChips />);
+    render(<VotingPowerChips isLoading isShown />);
     expect(screen.getByRole("progressbar")).toBeInTheDocument();
   });
 
   it("displays formatted ADA amount when data is available and not loading", () => {
-    mockUseCardano.mockReturnValue({
-      isEnableLoading: null,
-    } as ReturnType<typeof Context.useCardano>);
-    mockUseGetDRepVotingPowerQuery.mockReturnValue({
-      dRepVotingPower: 1000,
-    } as ReturnType<typeof Hooks.useGetDRepVotingPowerQuery>);
     mockUseScreenDimension.mockReturnValue({
       isMobile: false,
       screenWidth: 1024,
@@ -52,22 +30,13 @@ describe("VotingPowerChips", () => {
     mockUseTranslation.mockReturnValue({
       t: (key: string) => key,
     } as ReturnType<typeof Hooks.useTranslation>);
-    mockUseGetVoterInfo.mockReturnValue({
-      voter: { isRegisteredAsDRep: true },
-    } as ReturnType<typeof Hooks.useGetVoterInfo>);
     mockCorrectAdaFormat.mockReturnValue(1000);
 
-    render(<VotingPowerChips />);
+    render(<VotingPowerChips isShown />);
     expect(screen.getByText(/₳ 1000/)).toBeInTheDocument();
   });
 
   it("displays the tooltip correctly for non-mobile DRep registered users", async () => {
-    mockUseCardano.mockReturnValue({
-      isEnableLoading: null,
-    } as ReturnType<typeof Context.useCardano>);
-    mockUseGetDRepVotingPowerQuery.mockReturnValue({
-      dRepVotingPower: 1000,
-    } as ReturnType<typeof Hooks.useGetDRepVotingPowerQuery>);
     mockUseScreenDimension.mockReturnValue({
       isMobile: false,
       screenWidth: 800,
@@ -75,12 +44,10 @@ describe("VotingPowerChips", () => {
     mockUseTranslation.mockReturnValue({
       t: (key: string) => key,
     } as ReturnType<typeof Hooks.useTranslation>);
-    mockUseGetVoterInfo.mockReturnValue({
-      voter: { isRegisteredAsDRep: true },
-    } as ReturnType<typeof Hooks.useGetVoterInfo>);
+
     mockCorrectAdaFormat.mockReturnValue(1000);
 
-    render(<VotingPowerChips />);
+    render(<VotingPowerChips isShown />);
 
     const icon = screen.getByTestId("InfoOutlinedIcon");
     fireEvent.mouseOver(icon);