+++ govtool/metadata-validation/src/enums/index.ts
@@ -1,1 +1,2 @@
 export * from './ValidationError';
+export * from './LoggerMessage';